% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpStrmPath.R
\name{getUpStrmPath}
\alias{getUpStrmPath}
\title{Upstream path}
\usage{
getUpStrmPath(site, configurationFile, include.MMR.sites = F,
  rootSite = "GRA")
}
\arguments{
\item{site}{character value of site code}

\item{configurationFile}{data.frame with column called SiteID with site codes, a column called RKM with the river mask codes, and a column called SiteType with values of either \code{INT} for interrogation sites or \code{MMR} for mark/recapture/recovery sites}

\item{include.MMR.sites}{\code{TRUE/FALSE} should MMR sites be included. Default is \code{FALSE}}

\item{rootSite}{character value of the site code for the root site for which the path should be traced from}
}
\description{
For a given site, uses the configuration file to determine which path (through which sites) must be taken from the \code{rootSite} to reach that site.
}
\author{
Kevin See
}
